Added new types, remove unused imports
